requirements.txt
beautifulsoup4
requests
polar